requirements.txt
pymupdf
langchain
sentence-transformers
faiss-cpu
numpy
requests